Finalizacion del tp3-sql (Laboratorio 3), tiene un error en el cual no se puede guardar los precios_publicos con decimales.

+++ Laboratorio 3/Unidad 1/Codigo/tp3-sql/video 3/Productos.cs
@@ -12,13 +12,13 @@ namespace video_3
         private string codigo;
         private string nombre;
         private string descripcion;
-        private double precio_publico;
+        private float precio_publico;
         private int existencia;
 
         public int Id { get => id; set => id = value; }
         public string Nombre { get => nombre; set => nombre = value; }
         public string Descripcion { get => descripcion; set => descripcion = value; }
-        public double Precio_publico { get => precio_publico; set => precio_publico = value; }
+        public float Precio_publico { get => precio_publico; set => precio_publico = value; }
         public int Existencia { get => existencia; set => existencia = value; }
         public string Codigo { get => codigo; set => codigo = value; }
     }
